First pass of transitioning over to object reference creation
