+++ examples/Cassie/kinematic_centroidal_mpc/BUILD.bazel
@@ -59,11 +59,12 @@ cc_library(
 
 cc_library(
     name = "kinematic_centroidal_reference_generator",
-    srcs = ["reference_generator.cc"],
-    hdrs = ["reference_generator.h"],
+    srcs = ["cassie_reference_utils.cc"],
+    hdrs = ["cassie_reference_utils.h"],
     deps = [
         "//examples/Cassie:cassie_utils",
         "//common",
+        "//systems/controllers/kinematic_centroidal_mpc:kinematic_centroidal_reference_generator",
         "//multibody:visualization_utils",
         "@drake//:drake_shared_library",
     ],
